% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\docType{data}
\name{RealData.ROT}
\alias{RealData.ROT}
\title{RealData.ROT: A simulated dataset based on a pre-specified time-varying coefficients Cox model.}
\format{
An object of class \code{data.frame} with 444 rows and 8 columns.
}
\usage{
RealData.ROT
}
\description{
RealData.ROT
}
\examples{

\donttest{
# An example of illustrating this dataset can be found in the help page of our
#    function \code{SurvMA.Fit()} by typing \code{?SurvMA.Fit()}.

# It was originally extracted from the dataset rotteram in R package survival
# The specific extractions can be done using the following R commands
library(survival)
RealData.ROT <- na.omit(rotterdam[
  rotterdam$year \%in\% c(1992,1993),-c(1,2,5,6,7,12,13)
])
rownames(RealData.ROT) <- NULL
colnames(RealData.ROT)[c(7,8)] <- c("time","delta")
}

}
\keyword{datasets}
