% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\docType{data}
\name{SimData.APL}
\alias{SimData.APL}
\title{SimData.APL: A simulated dataset based on a pre-specified partly linear additive Cox model.}
\format{
An object of class \code{data.frame} with 200 rows and 9 columns.
}
\usage{
SimData.APL
}
\description{
SimData.APL: A simulated dataset based on a pre-specified partly linear additive Cox model.
}
\examples{

\donttest{
# An example of illustrating this dataset can be found in the help page of our
#    function \code{SurvMA.Fit()} by typing \code{?SurvMA.Fit()}.
}

}
\keyword{datasets}
