% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\docType{data}
\name{SimData.TVC}
\alias{SimData.TVC}
\title{SimData.TVC: A simulated dataset based on a pre-specified time-varying coefficients Cox model.}
\format{
An object of class \code{data.frame} with 150 rows and 8 columns.
}
\usage{
SimData.TVC
}
\description{
SimData.TVC: A simulated dataset based on a pre-specified time-varying coefficients Cox model.
}
\examples{

\donttest{
# An example of illustrating this dataset can be found in the help page of our
#    function \code{SurvMA.Fit()} by typing \code{?SurvMA.Fit()}.
}

}
\keyword{datasets}
