% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{SurvMA.Predict}
\alias{SurvMA.Predict}
\title{Model averaging prediction of personalized survival probabilities (prediction)}
\usage{
SurvMA.Predict(object, covariates, times)
}
\arguments{
\item{object}{a list of all outputted results from another main function named \code{SurvMA.Fit()}.}

\item{covariates}{is a data.frame with rows representing individuals and columns containing the necessary covariates used in the \code{formula} argument of \code{SurvMA.Fit()}.}

\item{times}{specifies the time points at which survival probabilities will be calculated.}
}
\value{
A list of fitted results that contain, for example, predicted values of personalized survival probabilities.
}
\description{
Model averaging prediction of personalized survival probabilities (prediction)
}
\details{
This is a function used to conduct model averaging prediction (prediction) of personalized survival probabilities.
For preliminary model fitting process, see another function \code{SurvMA.Fit()}.
}
\examples{

\donttest{

# Examples of illustrating the usages of this function can be found in the help
#    page of our another function \code{SurvMA.Fit()} by typing \code{?SurvMA.Fit()}.

}

}
