% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\docType{package}
\name{SurvMA}
\alias{SurvMA}
\title{SurvMA: Model Averaging Prediction of Personalized Survival Probabilities using R Package SurvMA.}
\description{
This package provides model averaging-based approaches that can be used to predict personalized survival probabilities.
}
