% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\docType{package}
\name{SurvMA}
\alias{SurvMA}
\title{SurvMA: Model Averaging Prediction of Personalized Survival Probabilities using R Package SurvMA.}
\description{
Provide model averaging-based approaches that can be used to predict personalized survival probabilities.
The key underlying idea is to approximate the conditional survival function using a weighted average of multiple candidate models.
Two scenarios of candidate models are allowed: (Scenario 1) partial linear Cox model and (Scenario 2) time-varying coefficient Cox model.
A reference of the underlying methods is Li and Wang (2023) \url{doi:10.1016/j.csda.2023.107759}.
}
